requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
attrs==24.2.0
audioread==3.0.1
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
decorator==5.1.1
dill==0.3.9
frozenlist==1.4.1
gevent==24.10.3
geventhttpclient==2.0.2
greenlet==3.1.1
grpcio==1.67.0
idna==3.10
importlib_metadata==8.5.0
joblib==1.4.2
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
markdown-it-py==3.0.0
mdurl==0.1.2
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
numba==0.60.0
numpy==1.26.4
nvidia-pytriton==0.5.12
packaging==24.1
pandarallel==1.6.5
pandas==2.2.3
platformdirs==4.3.6
pooch==1.8.2
propcache==0.2.0
protobuf==4.25.5
psutil==6.1.0
pycparser==2.22
pydub==0.25.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-rapidjson==1.20
pytz==2024.2
pyzmq==26.2.0
requests==2.32.3
rich==13.9.2
scikit-learn==1.5.2
scipy==1.14.1
sh==2.1.0
shellingham==1.5.4
six==1.16.0
soundfile==0.12.1
soxr==0.5.0.post1
threadpoolctl==3.5.0
tqdm==4.67.0
tritonclient==2.50.0
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wrapt==1.16.0
yarl==1.15.5
zipp==3.20.2
zope.event==5.0
zope.interface==7.1.0
